+++ build.gradle
@@ -99,25 +99,25 @@ subprojects {
 
 project(':micro-api') {
     dependencies {
-        compile project(':micro-common') , project(':micro-interfaces') , project(':micro-rdb')
+        implementation project(':micro-common') , project(':micro-interfaces') , project(':micro-rdb')
     }
 }
 
 project(':micro-common') {
     dependencies {
-        compile project(':micro-interfaces')
+        implementation project(':micro-interfaces')
     }
 }
 
 project(':micro-interfaces') {
     dependencies {
-        compile project(':micro-validation')
+        implementation project(':micro-validation')
     }
 }
 
 project(':micro-rdb') {
     dependencies {
-        compile project(':micro-common')
+        implementation project(':micro-common') , project(':micro-interfaces')
     }
 }
