+++ micro-api/src/main/java/com/mabubu0203/sudoku/api/service/UpdateService.java
@@ -6,8 +6,6 @@ import org.springframework.http.ResponseEntity;
 public interface UpdateService {
 
     /**
-     *
-     *
      * @author uratamanabu
      * @since 1.0
      */
