+++ micro-api/src/main/java/com/mabubu0203/sudoku/api/service/impl/SearchServiceImpl.java
@@ -4,7 +4,7 @@ import com.mabubu0203.sudoku.api.service.SearchService;
 import com.mabubu0203.sudoku.interfaces.NumberPlaceBean;
 import com.mabubu0203.sudoku.interfaces.SearchConditionBean;
 import com.mabubu0203.sudoku.interfaces.response.ScoreResponseBean;
-import com.mabubu0203.sudoku.interfaces.response.SearchResultBean;
+import com.mabubu0203.sudoku.interfaces.response.SearchSudokuRecordResponseBean;
 import com.mabubu0203.sudoku.rdb.domain.AnswerInfoTbl;
 import com.mabubu0203.sudoku.rdb.domain.ScoreInfoTbl;
 import com.mabubu0203.sudoku.rdb.service.AnswerInfoService;
@@ -12,13 +12,13 @@ import com.mabubu0203.sudoku.rdb.service.ScoreInfoService;
 import lombok.extern.slf4j.Slf4j;
 import ma.glasnost.orika.MapperFacade;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 @Slf4j
@@ -59,8 +59,9 @@ public class SearchServiceImpl implements SearchService {
         if (page != null && page.hasContent()) {
             Page modiftyPage = convertJacksonFile(page);
             SearchSudokuRecordResponseBean response = new SearchSudokuRecordResponseBean();
-            response.setPage(modiftyPage);
-            response.setPh(new PagenationHelper(modiftyPage));
+            // TODO:移植中
+//            response.setPage(modiftyPage);
+//            response.setPh(new PagenationHelper(modiftyPage));
             return new ResponseEntity<>(response, HttpStatus.OK);
         } else {
             SearchSudokuRecordResponseBean response = new SearchSudokuRecordResponseBean();
@@ -71,7 +72,7 @@ public class SearchServiceImpl implements SearchService {
     @Override
     public ResponseEntity<NumberPlaceBean> getNumberPlaceDetail(final int type) {
         AnswerInfoTbl answerInfoTbl = answerInfoService.findByType(type);
-        NumberPlaceBean numberPlaceBean = SudokuUtils.answerInfoTblConvertBean(answerInfoTbl);
+        NumberPlaceBean numberPlaceBean = answerInfoService.answerInfoTblConvertBean(answerInfoTbl);
         return new ResponseEntity<>(numberPlaceBean, HttpStatus.OK);
     }
 
@@ -79,7 +80,7 @@ public class SearchServiceImpl implements SearchService {
     public ResponseEntity<NumberPlaceBean> getNumberPlaceDetail(
             final int type, final String keyHash) {
         AnswerInfoTbl answerInfoTbl = answerInfoService.findByTypeAndKeyHash(type, keyHash);
-        NumberPlaceBean numberPlaceBean = SudokuUtils.answerInfoTblConvertBean(answerInfoTbl);
+        NumberPlaceBean numberPlaceBean = answerInfoService.answerInfoTblConvertBean(answerInfoTbl);
         return new ResponseEntity<>(numberPlaceBean, HttpStatus.OK);
     }
 
@@ -94,22 +95,24 @@ public class SearchServiceImpl implements SearchService {
 
     private Page convertJacksonFile(final Page<AnswerInfoTbl> page) {
 
-        Pageable pageable =
-                new PageRequest(page.getNumber(), page.getSize(), new OrderBySource("NoAsc").toSort());
-        List<AnswerInfoTbl> content = page.getContent();
-        List<SearchResultBean> modifyContent = new ArrayList<>();
-        for (AnswerInfoTbl record : content) {
-            SearchResultBean bean = new SearchResultBean();
-            bean.setNo(record.getNo());
-            bean.setType(record.getType());
-            bean.setKeyHash(record.getKeyHash());
-            ScoreInfoTbl scoreInfoTbl = record.getScoreInfoTbl();
-            bean.setName(scoreInfoTbl.getName());
-            bean.setScore(scoreInfoTbl.getScore());
-            bean.setMemo(scoreInfoTbl.getMemo());
-            modifyContent.add(bean);
-        }
-        Page result = new PageImplBean(modifyContent, pageable, page.getTotalElements());
+        // TODO:移植中
+//        Pageable pageable =
+//                new PageRequest(page.getNumber(), page.getSize(), new OrderBySource("NoAsc").toSort());
+//        List<AnswerInfoTbl> content = page.getContent();
+//        List<SearchResultBean> modifyContent = new ArrayList<>();
+//        for (AnswerInfoTbl record : content) {
+//            SearchResultBean bean = new SearchResultBean();
+//            bean.setNo(record.getNo());
+//            bean.setType(record.getType());
+//            bean.setKeyHash(record.getKeyHash());
+//            ScoreInfoTbl scoreInfoTbl = record.getScoreInfoTbl();
+//            bean.setName(scoreInfoTbl.getName());
+//            bean.setScore(scoreInfoTbl.getScore());
+//            bean.setMemo(scoreInfoTbl.getMemo());
+//            modifyContent.add(bean);
+//        }
+//        Page result = new PageImplBean(modifyContent, pageable, page.getTotalElements());
+        Page result = new PageImpl<AnswerInfoTbl>(Collections.emptyList());
         return result;
     }
