+++ micro-common/src/main/java/com/mabubu0203/sudoku/constants/PathParameterConstants.java
@@ -6,6 +6,6 @@ import static lombok.AccessLevel.PRIVATE;
 
 @NoArgsConstructor(access = PRIVATE)
 public class PathParameterConstants {
-  public static final String PATH_TYPE = "{type}";
-  public static final String PATH_TYPEANSWER_KEY = "{answerKey}";
+    public static final String PATH_TYPE = "{type}";
+    public static final String PATH_TYPEANSWER_KEY = "{answerKey}";
 }
