+++ micro-common/src/main/java/com/mabubu0203/sudoku/constants/RestUrlConstants.java
@@ -6,10 +6,10 @@ import static lombok.AccessLevel.PRIVATE;
 
 @NoArgsConstructor(access = PRIVATE)
 public class RestUrlConstants {
-  public static final String URL_SEARCH_MASTER = "searchMaster";
-  public static final String URL_CREATE_MASTER = "createMaster";
-  public static final String URL_UPDATE_MASTER = "updateMaster";
-  public static final String URL_GENERATE = "generate";
-  public static final String URL_SUDOKU = "sudoku";
-  public static final String URL_SCORE = "score";
+    public static final String URL_SEARCH_MASTER = "searchMaster";
+    public static final String URL_CREATE_MASTER = "createMaster";
+    public static final String URL_UPDATE_MASTER = "updateMaster";
+    public static final String URL_GENERATE = "generate";
+    public static final String URL_SUDOKU = "sudoku";
+    public static final String URL_SCORE = "score";
 }
