+++ micro-common/src/main/java/com/mabubu0203/sudoku/controller/RestBaseController.java
@@ -3,4 +3,5 @@ package com.mabubu0203.sudoku.controller;
 import lombok.ToString;
 
 @ToString
-public abstract class RestBaseController {}
+public abstract class RestBaseController {
+}
