+++ micro-common/src/main/java/com/mabubu0203/sudoku/controller/WebBaseController.java
@@ -12,4 +12,5 @@ import lombok.ToString;
  * @since 1.0
  */
 @ToString
-public abstract class WebBaseController {}
+public abstract class WebBaseController {
+}
