+++ micro-common/src/main/java/com/mabubu0203/sudoku/enums/Type.java
@@ -10,33 +10,40 @@ import lombok.Getter;
  * @since 1.0
  */
 public enum Type {
-  /** size:4 */
-  SIZE4(4),
+    /**
+     * size:4
+     */
+    SIZE4(4),
 
-  /** size:9 */
-  SIZE9(9);
+    /**
+     * size:9
+     */
+    SIZE9(9);
 
-  @Getter private int size;
+    @Getter
+    private int size;
 
-  /** コンストラクタ */
-  Type(int size) {
-    this.size = size;
-  }
-
-  /**
-   * 列挙型を返却します。
-   *
-   * @return 数独のサイズ
-   * @author uratamanabu
-   * @version 1.0
-   * @since 1.0
-   */
-  public static Type getType(final int size) {
-    for (Type type : Type.values()) {
-      if (size == type.getSize()) {
-        return type;
-      }
+    /**
+     * コンストラクタ
+     */
+    Type(int size) {
+        this.size = size;
     }
-    return null;
-  }
+
+    /**
+     * 列挙型を返却します。
+     *
+     * @return 数独のサイズ
+     * @author uratamanabu
+     * @version 1.0
+     * @since 1.0
+     */
+    public static Type getType(final int size) {
+        for (Type type : Type.values()) {
+            if (size == type.getSize()) {
+                return type;
+            }
+        }
+        return null;
+    }
 }
