+++ micro-common/src/main/java/com/mabubu0203/sudoku/exception/SudokuApplicationException.java
@@ -10,5 +10,5 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class SudokuApplicationException extends RuntimeException {
 
-  private static final long serialVersionUID = -4376900603436217304L;
+    private static final long serialVersionUID = -4376900603436217304L;
 }
