+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/interfaces/NumberPlaceDefine.java
@@ -9,96 +9,96 @@ package com.mabubu0203.sudoku.interfaces;
  */
 public interface NumberPlaceDefine {
 
-  int type = 0;
+    int type = 0;
 
-  int a1 = 0;
-  int a2 = 0;
-  int a3 = 0;
-  int a4 = 0;
-  int a5 = 0;
-  int a6 = 0;
-  int a7 = 0;
-  int a8 = 0;
-  int a9 = 0;
+    int a1 = 0;
+    int a2 = 0;
+    int a3 = 0;
+    int a4 = 0;
+    int a5 = 0;
+    int a6 = 0;
+    int a7 = 0;
+    int a8 = 0;
+    int a9 = 0;
 
-  int b1 = 0;
-  int b2 = 0;
-  int b3 = 0;
-  int b4 = 0;
-  int b5 = 0;
-  int b6 = 0;
-  int b7 = 0;
-  int b8 = 0;
-  int b9 = 0;
+    int b1 = 0;
+    int b2 = 0;
+    int b3 = 0;
+    int b4 = 0;
+    int b5 = 0;
+    int b6 = 0;
+    int b7 = 0;
+    int b8 = 0;
+    int b9 = 0;
 
-  int c1 = 0;
-  int c2 = 0;
-  int c3 = 0;
-  int c4 = 0;
-  int c5 = 0;
-  int c6 = 0;
-  int c7 = 0;
-  int c8 = 0;
-  int c9 = 0;
+    int c1 = 0;
+    int c2 = 0;
+    int c3 = 0;
+    int c4 = 0;
+    int c5 = 0;
+    int c6 = 0;
+    int c7 = 0;
+    int c8 = 0;
+    int c9 = 0;
 
-  int d1 = 0;
-  int d2 = 0;
-  int d3 = 0;
-  int d4 = 0;
-  int d5 = 0;
-  int d6 = 0;
-  int d7 = 0;
-  int d8 = 0;
-  int d9 = 0;
+    int d1 = 0;
+    int d2 = 0;
+    int d3 = 0;
+    int d4 = 0;
+    int d5 = 0;
+    int d6 = 0;
+    int d7 = 0;
+    int d8 = 0;
+    int d9 = 0;
 
-  int e1 = 0;
-  int e2 = 0;
-  int e3 = 0;
-  int e4 = 0;
-  int e5 = 0;
-  int e6 = 0;
-  int e7 = 0;
-  int e8 = 0;
-  int e9 = 0;
+    int e1 = 0;
+    int e2 = 0;
+    int e3 = 0;
+    int e4 = 0;
+    int e5 = 0;
+    int e6 = 0;
+    int e7 = 0;
+    int e8 = 0;
+    int e9 = 0;
 
-  int f1 = 0;
-  int f2 = 0;
-  int f3 = 0;
-  int f4 = 0;
-  int f5 = 0;
-  int f6 = 0;
-  int f7 = 0;
-  int f8 = 0;
-  int f9 = 0;
+    int f1 = 0;
+    int f2 = 0;
+    int f3 = 0;
+    int f4 = 0;
+    int f5 = 0;
+    int f6 = 0;
+    int f7 = 0;
+    int f8 = 0;
+    int f9 = 0;
 
-  int g1 = 0;
-  int g2 = 0;
-  int g3 = 0;
-  int g4 = 0;
-  int g5 = 0;
-  int g6 = 0;
-  int g7 = 0;
-  int g8 = 0;
-  int g9 = 0;
+    int g1 = 0;
+    int g2 = 0;
+    int g3 = 0;
+    int g4 = 0;
+    int g5 = 0;
+    int g6 = 0;
+    int g7 = 0;
+    int g8 = 0;
+    int g9 = 0;
 
-  int h1 = 0;
-  int h2 = 0;
-  int h3 = 0;
-  int h4 = 0;
-  int h5 = 0;
-  int h6 = 0;
-  int h7 = 0;
-  int h8 = 0;
-  int h9 = 0;
+    int h1 = 0;
+    int h2 = 0;
+    int h3 = 0;
+    int h4 = 0;
+    int h5 = 0;
+    int h6 = 0;
+    int h7 = 0;
+    int h8 = 0;
+    int h9 = 0;
 
-  int i1 = 0;
-  int i2 = 0;
-  int i3 = 0;
-  int i4 = 0;
-  int i5 = 0;
-  int i6 = 0;
-  int i7 = 0;
-  int i8 = 0;
-  int i9 = 0;
+    int i1 = 0;
+    int i2 = 0;
+    int i3 = 0;
+    int i4 = 0;
+    int i5 = 0;
+    int i6 = 0;
+    int i7 = 0;
+    int i8 = 0;
+    int i9 = 0;
 
 }
