+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/interfaces/request/ResisterSudokuRecordRequestBean.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 public class ResisterSudokuRecordRequestBean {
 
-  private NumberPlaceBean numberPlaceBean;
+    private NumberPlaceBean numberPlaceBean;
 }
