+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/interfaces/response/SearchResultBean.java
@@ -8,13 +8,13 @@ import java.time.LocalDateTime;
 @Data
 public class SearchResultBean implements Serializable {
 
-  private long no;
-  private int type;
-  private String answerKey;
-  private String keyHash;
-  private String name;
-  private int score;
-  private String memo;
-  private LocalDateTime createDate;
+    private long no;
+    private int type;
+    private String answerKey;
+    private String keyHash;
+    private String name;
+    private int score;
+    private String memo;
+    private LocalDateTime createDate;
 
 }
