+++ micro-interfaces/src/main/java/com/mabubu0203/sudoku/interfaces/response/SearchSudokuRecordResponseBean.java
@@ -1,0 +1,16 @@
+package com.mabubu0203.sudoku.interfaces.response;
+
+import lombok.Data;
+
+/**
+ * Created by uratamanabu on 2017/07/04.
+ */
+@Data
+@Deprecated
+public class SearchSudokuRecordResponseBean {
+
+//  private PagenationHelper ph;
+//
+//  @JsonTypeInfo(use = Id.NAME, defaultImpl = PageImplBean.class)
+//  private Page page;
+}
