+++ micro-rdb/build.gradle
@@ -11,6 +11,12 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-validation")
     testCompile("org.springframework.boot:spring-boot-starter-test")
 
+    // EclipseCollections
+    implementation("org.eclipse.collections:eclipse-collections:$ecVersion")
+    implementation("org.eclipse.collections:eclipse-collections-api:$ecVersion")
+    testCompile("org.eclipse.collections:eclipse-collections-testutils:$ecVersion")
+    implementation("org.eclipse.collections:eclipse-collections-forkjoin:$ecVersion")
+
     // Flyway
     implementation("org.flywaydb:flyway-core")
 
@@ -19,10 +25,10 @@ dependencies {
     compileOnly("org.projectlombok:lombok")
 
     // JDBC Drivers
-    compile("mysql:mysql-connector-java")
+    implementation("mysql:mysql-connector-java")
 
     // ModelMapper
-    compile("com.github.jmnarloch:modelmapper-spring-boot-starter:1.1.0")
+    implementation("com.github.jmnarloch:modelmapper-spring-boot-starter:1.1.0")
 
 }
