+++ micro-rdb/src/main/java/com/mabubu0203/sudoku/rdb/service/impl/AnswerInfoServiceImpl.java
@@ -1,5 +1,7 @@
 package com.mabubu0203.sudoku.rdb.service.impl;
 
+import com.mabubu0203.sudoku.enums.Type;
+import com.mabubu0203.sudoku.exception.SudokuApplicationException;
 import com.mabubu0203.sudoku.interfaces.NumberPlaceBean;
 import com.mabubu0203.sudoku.interfaces.SearchConditionBean;
 import com.mabubu0203.sudoku.rdb.domain.AnswerInfoTbl;
@@ -7,6 +9,8 @@ import com.mabubu0203.sudoku.rdb.repository.AnswerInfoRepository;
 import com.mabubu0203.sudoku.rdb.service.AnswerInfoService;
 import com.mabubu0203.sudoku.rdb.specification.AnswerInfoSpecifications;
 import com.mabubu0203.sudoku.rdb.specification.ScoreInfoSpecifications;
+import com.mabubu0203.sudoku.utils.ESListWrapUtils;
+import com.mabubu0203.sudoku.utils.SudokuUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
@@ -17,6 +21,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.ListIterator;
 
 /**
  * ANSWER_INFO_TBLへのサービスクラスです。 このクラスを経由してCRUD操作を実行してください。
@@ -102,4 +107,32 @@ public class AnswerInfoServiceImpl implements AnswerInfoService {
         Page<AnswerInfoTbl> page = answerInfoRepository.findAll(answerSpecification, pageable);
         return page;
     }
+
+    @Override
+    public NumberPlaceBean answerInfoTblConvertBean(AnswerInfoTbl answerInfoTbl)
+            throws SudokuApplicationException {
+        NumberPlaceBean numberPlaceBean = new ModelMapper().map(answerInfoTbl, NumberPlaceBean.class);
+        String answerKey = numberPlaceBean.getAnswerKey();
+        String[] valueArray = answerKey.split("");
+        int size = 0;
+        Type type = Type.getType(numberPlaceBean.getType());
+        if (type == null) {
+            throw new SudokuApplicationException();
+        } else {
+            size = type.getSize();
+        }
+        ListIterator<String> itr = ESListWrapUtils.createCells(size, 0).listIterator();
+        try {
+            for (String value : valueArray) {
+                SudokuUtils.setCell(numberPlaceBean, itr.next(), Integer.valueOf(value));
+            }
+        } catch (SudokuApplicationException e) {
+            e.printStackTrace();
+            log.error("やらかしています。");
+            throw new SudokuApplicationException();
+        }
+        return numberPlaceBean;
+    }
+
+
 }
