+++ micro-rdb/src/main/java/com/mabubu0203/sudoku/rdb/specification/AnswerInfoSpecifications.java
@@ -19,74 +19,71 @@ import static lombok.AccessLevel.PRIVATE;
 @NoArgsConstructor(access = PRIVATE)
 public class AnswerInfoSpecifications {
 
-  private static final String NO = "no";
-  private static final String TYPE = "type";
-  private static final String KEYHASH = "keyHash";
+    private static final String NO = "no";
+    private static final String TYPE = "type";
+    private static final String KEYHASH = "keyHash";
 
-  /**
-   *
-   * @author uratamanabu
-   * @since 1.0
-   */
-  public static Specification<AnswerInfoTbl> typeContains(int selectType) {
-    return (root, query, builder) -> builder.equal(root.get(TYPE), selectType);
-  }
-
-  /**
-   *
-   * @author uratamanabu
-   * @since 1.0
-   */
-  public static Specification<AnswerInfoTbl> noContains(long no, int selectorNo) {
-    if (no == 0) {
-      return null;
-    } else {
-      Selector selector = Selector.getSelector(selectorNo);
-      if (selector == null) {
-        return (root, query, builder) -> builder.equal(root.get(NO), no);
-      }
-      switch (selector) {
-        case PERFECT_MATCH:
-          return (root, query, builder) -> builder.equal(root.get(NO), no);
-        case AROUND5:
-          long fromNo = no - 5 < 0 ? 0 : no - 5;
-          long endNo = no + 5 > 9999999999L ? 9999999999L : no + 5;
-          return (root, query, builder) -> builder.between(root.get(NO), fromNo, endNo);
-        case MORE_BIG:
-          return (root, query, builder) -> builder.greaterThan(root.get(NO), no);
-        case MORE_SMALL:
-          return (root, query, builder) -> builder.lessThan(root.get(NO), no);
-        default:
-          return (root, query, builder) -> builder.equal(root.get(NO), no);
-      }
+    /**
+     * @author uratamanabu
+     * @since 1.0
+     */
+    public static Specification<AnswerInfoTbl> typeContains(int selectType) {
+        return (root, query, builder) -> builder.equal(root.get(TYPE), selectType);
     }
-  }
 
-  /**
-   *
-   * @author uratamanabu
-   * @since 1.0
-   */
-  public static Specification<AnswerInfoTbl> keyHashContains(String keyHash, int selectorKeyHash) {
-    if (keyHash == null) {
-      return null;
-    } else {
-      Selector selector = Selector.getSelector(selectorKeyHash);
-      if (selector == null) {
-        return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
-      }
-      switch (selector) {
-        case PERFECT_MATCH:
-          return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
-        case FORWARD_MATCH:
-          return (root, query, builder) -> builder.like(root.get(KEYHASH), keyHash + '%');
-        case BACKWARD_MATCH:
-          return (root, query, builder) -> builder.like(root.get(KEYHASH), '%' + keyHash);
-        case PARTIAL_MATCH:
-          return (root, query, builder) -> builder.like(root.get(KEYHASH), '%' + keyHash + '%');
-        default:
-          return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
-      }
+    /**
+     * @author uratamanabu
+     * @since 1.0
+     */
+    public static Specification<AnswerInfoTbl> noContains(long no, int selectorNo) {
+        if (no == 0) {
+            return null;
+        } else {
+            Selector selector = Selector.getSelector(selectorNo);
+            if (selector == null) {
+                return (root, query, builder) -> builder.equal(root.get(NO), no);
+            }
+            switch (selector) {
+                case PERFECT_MATCH:
+                    return (root, query, builder) -> builder.equal(root.get(NO), no);
+                case AROUND5:
+                    long fromNo = no - 5 < 0 ? 0 : no - 5;
+                    long endNo = no + 5 > 9999999999L ? 9999999999L : no + 5;
+                    return (root, query, builder) -> builder.between(root.get(NO), fromNo, endNo);
+                case MORE_BIG:
+                    return (root, query, builder) -> builder.greaterThan(root.get(NO), no);
+                case MORE_SMALL:
+                    return (root, query, builder) -> builder.lessThan(root.get(NO), no);
+                default:
+                    return (root, query, builder) -> builder.equal(root.get(NO), no);
+            }
+        }
     }
-  }
+
+    /**
+     * @author uratamanabu
+     * @since 1.0
+     */
+    public static Specification<AnswerInfoTbl> keyHashContains(String keyHash, int selectorKeyHash) {
+        if (keyHash == null) {
+            return null;
+        } else {
+            Selector selector = Selector.getSelector(selectorKeyHash);
+            if (selector == null) {
+                return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
+            }
+            switch (selector) {
+                case PERFECT_MATCH:
+                    return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
+                case FORWARD_MATCH:
+                    return (root, query, builder) -> builder.like(root.get(KEYHASH), keyHash + '%');
+                case BACKWARD_MATCH:
+                    return (root, query, builder) -> builder.like(root.get(KEYHASH), '%' + keyHash);
+                case PARTIAL_MATCH:
+                    return (root, query, builder) -> builder.like(root.get(KEYHASH), '%' + keyHash + '%');
+                default:
+                    return (root, query, builder) -> builder.equal(root.get(KEYHASH), keyHash);
+            }
+        }
+    }
 }
