+++ micro-validation/src/main/java/com/mabubu0203/sudoku/validator/constraint/KeyHash.java
@@ -2,7 +2,6 @@ package com.mabubu0203.sudoku.validator.constraint;
 
 
 import javax.validation.Constraint;
-
 import javax.validation.Payload;
 import javax.validation.ReportAsSingleViolation;
 import javax.validation.constraints.Size;
@@ -20,9 +19,9 @@ import java.lang.annotation.Target;
 @Size(max = 64)
 public @interface KeyHash {
 
-  String message() default "KeyHash is invalid";
+    String message() default "KeyHash is invalid";
 
-  Class<?>[] groups() default {};
+    Class<?>[] groups() default {};
 
-  Class<? extends Payload>[] payload() default {};
+    Class<? extends Payload>[] payload() default {};
 }
